+++ src/test/se/jbee/inject/bind/TestPresetModuleBinds.java
@@ -4,7 +4,7 @@ import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static se.jbee.inject.Name.named;
-import static se.jbee.inject.container.Typecast.listTypeOf;
+import static se.jbee.inject.container.Cast.listTypeOf;
 
 import java.util.List;
 import java.util.Properties;